Allow customization of plugin creation in tests
